+++ frontend/android/app/src/main/java/com/pbl3/dashboard/MainActivity.java
@@ -1,0 +1,5 @@
+package com.pbl3.dashboard;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}